Moved help and tournament commands to respective cogs
